+++ ResourciaFrontend/src/app/layouts/app-layout/components/app-top-nav/app-top-nav.component.ts
@@ -1,12 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { NavigationEnd, Router, RouterLink } from '@angular/router';
+import { RouterLink } from '@angular/router';
 import { AuthService } from '../../../../core/auth/auth.service';
 import { ExploreMenuComponent } from './components/explore-menu/explore-menu.component';
 import { NavSearchComponent } from './components/nav-search/nav-search.component';
 import { BrandLogoComponent } from './components/brand-logo/brand-logo.component';
 import { AuthLinksComponent } from './components/auth-links/auth-links.component';
-import { JwtPayloadModel } from '../../../../shared/models/jwt-payload-model';
-import { jwtDecode } from 'jwt-decode';
+import { MeInfoModel } from '../../../../shared/models/me-info';
 import { LanguageSwitcherComponent } from './components/language-switcher/language-switcher.component';
 
 @Component({
@@ -16,63 +15,19 @@ import { LanguageSwitcherComponent } from './components/language-switcher/langua
   styleUrl: './app-top-nav.component.scss'
 })
 export class AppTopNavComponent implements OnInit {
-  user: JwtPayloadModel | null = null;
+  user: MeInfoModel | null = null;
   isMobileMenuOpen = false;
 
-  constructor(private authService: AuthService, private router: Router) {}
+  constructor(private authService: AuthService) {}
 
   ngOnInit(): void {
-    this.loadUser();
-
-    this.router.events.subscribe(e => {
-    if (e instanceof NavigationEnd) this.loadUser();
-  });
-
-  }
-
-  loadUser(): void {
-    const token = this.getToken();
-
-    if (!token) {
-      this.user = null;
-      return;
-    }
-
-    try {
-      const payload = jwtDecode<JwtPayloadModel>(token);
-
-      const now = Math.floor(Date.now() / 1000);
-      if (payload.exp && payload.exp <= now) {
-        this.user = null;
-        return;
-      }
-
-      const roles =
-      payload["http://schemas.microsoft.com/ws/2008/06/identity/claims/role"] ?? [];
-
-      const isAdmin = roles.includes('Admin');
-
-      console.log('JWT payload:', payload);
-      this.user = {
-        id: payload.id,
-        email: payload.email,
-        name: payload.name,
-        exp: payload.exp,
-        isAdmin,
-        "http://schemas.microsoft.com/ws/2008/06/identity/claims/role": roles
-      } satisfies JwtPayloadModel;
-    } catch (error) {
-      this.user = null;
-      console.error('Failed to decode JWT:', error);
-    }
+    this.authService.currentUser$.subscribe(user => {
+      this.user = user;
+    });
   }
 
   logout(): void {
     this.authService.logout();
-  }
-
-  private getToken(): string | null {
-    return localStorage.getItem('accessToken'); // or wherever you store it
   }
 
   toggleMobileMenu(): void {